Fixes signature verification for permit functions
Updates the domain separator calculation to correctly use the chain ID from the provider. This ensures that the signature verification in permit functions works as expected, especially in different network environments.

Also, conditionally configures the hardhat network based on the `USE_POLKAVM` environment variable. This allows to run tests with or without the polkavm enabled hardhat network.